+++ src/main/java/com/atlauncher/Launcher.java
@@ -95,7 +95,7 @@ public class Launcher {
 
             DialogManager.optionDialog().setTitle(GetText.tr("Unsupported Java Version"))
                     .setContent(new HTMLBuilder().center().text(GetText.tr(
-                            "You're using an unsupported version of Java. You need to upgrade your Java to at minimum Java 8 version 101.<br/><br/>The launcher will not start until you do this.<br/><br/>If you're seeing this message even after installing a newer version, you may need to uninstall the old version first.<br/><br/>Click ok to open the Java download page and close the launcher."))
+                            "You're using an unsupported version of Java. You need to upgrade your Java to at minimum Java 8 version 141.<br/><br/>The launcher will not start until you do this.<br/><br/>If you're seeing this message even after installing a newer version, you may need to uninstall the old version first.<br/><br/>Click ok to open the Java download page and close the launcher."))
                             .build())
                     .addOption(GetText.tr("Ok")).setType(DialogManager.ERROR).show();